test: make sure custom chain deserialization in python matches tycho

+++ crates/tycho-client-py/python/tycho_indexer_client/dto.py
@@ -55,8 +55,8 @@ class ChainTokenConfig(BaseModel):
 
 
 class TvlThresholds(BaseModel):
-    low: int
-    medium: int
+    low: float
+    medium: float
 
 
 class CustomChainConfig(BaseModel):
